+++ packages/connect-wallet-modal/src/components/Ledger/icons/LedgerImageErrorMobile.tsx
@@ -1,0 +1,98 @@
+import React, { ReactElement } from 'react';
+
+export const LedgerImageErrorMobile = (): ReactElement => {
+  return (
+    <div style={{ width: '48px', height: '260px' }}>
+      <svg
+        width="38"
+        height="258"
+        viewBox="0 0 38 258"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M1 6L1 204C1 206.761 3.23858 209 6 209L32 209C34.7614 209 37 206.761 37 204L37 6C37 3.23857 34.7614 1 32 1L6 1C3.23858 1 1 3.23857 1 6Z"
+          fill="#F6F7FB"
+          stroke="#34343D"
+          strokeWidth="2"
+        />
+        <path
+          d="M19 184C15.134 184 12 187.134 12 191C12 194.866 15.134 198 19 198C22.866 198 26 194.866 26 191C26 187.134 22.866 184 19 184Z"
+          fill="#F6F7FB"
+          stroke="#34343D"
+        />
+        <path
+          d="M5.5967 123L5.5967 167C5.5967 168.88 7.12042 170.403 9.00003 170.403L29.0001 170.403C30.8797 170.403 32.4034 168.88 32.4034 167L32.4034 123C32.4034 121.12 30.8797 119.597 29.0001 119.597L9.00003 119.597C7.12042 119.597 5.5967 121.12 5.5967 123Z"
+          fill="white"
+          stroke="#34343D"
+          strokeWidth="1.19333"
+        />
+        <path
+          d="M15 141L23 149"
+          stroke="#E14D4D"
+          strokeWidth="2"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        />
+        <path
+          d="M15 149L23 141"
+          stroke="#E14D4D"
+          strokeWidth="2"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        />
+        <path
+          d="M19 109C9.0589 109 1 100.941 1 91L1 5.03799C1 2.80799 2.80795 1 5.0382 1L32.9618 1C35.192 1 37 2.80799 37 5.03799L37 91C37 100.941 28.9411 109 19 109Z"
+          fill="white"
+          stroke="#34343D"
+          strokeWidth="2"
+        />
+        <path
+          d="M19 84C15.134 84 12 87.134 12 91C12 94.866 15.134 98 19 98C22.866 98 26 94.866 26 91C26 87.134 22.866 84 19 84Z"
+          fill="#F6F7FB"
+          stroke="#34343D"
+        />
+        <path
+          d="M16 223C14.3431 223 13 221.657 13 220L13 210.387C13 209.621 13.6208 209 14.3867 209L23.6133 209C24.3792 209 25 209.621 25 210.387L25 220C25 221.657 23.6569 223 22 223L16 223Z"
+          fill="white"
+          stroke="#34343D"
+          strokeWidth="2"
+        />
+        <path
+          d="M17.3867 232.5C16.3447 232.5 15.5 231.655 15.5 230.613L15.5 224.193C15.5 223.81 15.8104 223.5 16.1933 223.5L21.8067 223.5C22.1896 223.5 22.5 223.81 22.5 224.193L22.5 230.613C22.5 231.655 21.6553 232.5 20.6133 232.5L17.3867 232.5Z"
+          fill="white"
+          stroke="#34343D"
+        />
+        <path
+          d="M17.5 257.5L17.5 232.5L20.5 232.5L20.5 257.5L17.5 257.5Z"
+          fill="url(#paint0_linear_2_18)"
+          stroke="url(#paint1_linear_2_18)"
+        />
+        <defs>
+          <linearGradient
+            id="paint0_linear_2_18"
+            x1="19"
+            y1="231.649"
+            x2="19"
+            y2="258"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="white" />
+            <stop offset="1" stopColor="white" stopOpacity="0" />
+          </linearGradient>
+          <linearGradient
+            id="paint1_linear_2_18"
+            x1="21"
+            y1="232"
+            x2="21"
+            y2="255.892"
+            gradientUnits="userSpaceOnUse"
+          >
+            <stop stopColor="#34343D" />
+            <stop offset="1" stopColor="#34343D" stopOpacity="0" />
+          </linearGradient>
+        </defs>
+      </svg>
+    </div>
+  );
+};