+++ packages/connect-wallet-modal/src/components/Ledger/types.ts
@@ -1,0 +1,12 @@
+export type AccountRecord = {
+  index: string;
+  path: string;
+  pathTemplate: string;
+  address: string;
+  balance: number;
+  token: string;
+};
+
+export type AccountsStorage = {
+  [path: string]: { [accountIndex: string]: AccountRecord };
+};