+++ packages/connectors/ledger-connector/src/index.ts
@@ -1,2 +1,2 @@
-export { LedgerHQConnector } from './hid';
+export { LedgerHQConnector, LS_KEY_DERIVATION_PATH } from './hid';
 export * from './iframe';