+++ packages/web3-react/src/helpers/index.ts
@@ -2,4 +2,4 @@ export * from './providerDetectors';
 export * from './ua';
 export * from './openWindow';
 export { default as isUrl } from './isUrl';
-export { default as interceptLedgerError } from './interceptLedgerError';
+export * from './interceptLedgerError';