+++ packages/connectors/ledger-connector/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/ledger-connector
 
+## 1.1.0
+
+### Minor Changes
+
+- Rework direct Ledger hardware connection UI
+
 ## 1.0.1
 
 ### Patch Changes